requirements.txt
fastapi==0.135.1
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.12.5
pytest==9.0.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
httpx
pytest-asyncio
uvicorn